+++ src/database/constants.ts
@@ -1,5 +1,3 @@
-import { GuildConfig, RolesConfig } from './types';
-
 /**
  * Collection names
  */
@@ -15,14 +13,3 @@ export enum Config {
   GUILD = 'guild',
   ROLES = 'roles',
 }
-
-export const rolesConfig: RolesConfig = {
-  muted: '',
-};
-
-/**
- * Default value of the `guild` configuration
- */
-export const guildConfig: GuildConfig = {
-  commandPrefixes: ['-'],
-};
