+++ src/database/index.ts
@@ -1,3 +1,2 @@
 export { default as Database } from './Database';
 export { Collection, Config } from './constants';
-export { RolesConfig, GuildConfig } from './types';
