--- src/database/types.ts
@@ -1,7 +1,0 @@
-export type GuildConfig = {
-  commandPrefixes: string[];
-};
-
-export type RolesConfig = {
-  muted: string;
-};
